==> code_for_stats/code availability/10) gap predictions.Rmd <==
---
title: "9) gap predictions"
author: "Emmerson Wilson"
date: "2024-03-15"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra,
  brglm2,
  caret)

```

```{r}
# change this to whichever folder you downloaded the code to
setwd("~/Documents/Master's/C_forecasting_NL/code_for_stats/code availability")
```

# import training datasets
```{r}
# made in 4) gap glm.Rmd
GM_gap_env <- read_xlsx("data created/c_env_sites_scale_GM_gap.xlsx")
TN_gap_env <- read_xlsx("data created/c_env_sites_scale_TN_gap.xlsx")
```

# import scaled environmental rasters
```{r}
# made in 8) predictable space
GM_env_raster_masked_scaled_bm <- rast("data created/GM_env_raster_masked_scale.tif") 
TN_env_raster_masked_scaled_bm <- rast("data created/TN_env_raster_masked_scale.tif")
```

# import glm function
```{r}
source("2) glm function.R")
```

# predict

## GM
```{r}
# list of variables in best model from 5) c lm total.Rmd
GM_var_gap <- list("FHT", "EVIamp")

# run best model
GM_model_tot_gap <- firths_glm(GM_var_gap, df = GM_gap_env, c = "gap")

# predict
GM_predict_tot_gap <- terra::predict(GM_env_raster_masked_scaled_bm, GM_model_tot_gap)

# transform predictions
GM_predict_tot_gap_transform <- GM_predict_tot_gap
values(GM_predict_tot_gap_transform) <- 1 / (1 + exp(-values(GM_predict_tot_gap))) 

# export
writeRaster(GM_predict_tot_gap_transform, filename="data created/GM_predict_gap.tif", overwrite=TRUE)
writeRaster(GM_predict_tot_gap, filename="data created/GM_predict_gap1.tif", overwrite=TRUE)
```

## TN
```{r}
# list of variables in best model from 5) c lm total.Rmd
TN_var_gap <- list("FHT", "EVIamp")

# run best model
TN_model_tot_gap <- firths_glm(TN_var_gap, df = TN_gap_env, c = "gap")

# predict
TN_predict_tot_gap <- terra::predict(TN_env_raster_masked_scaled_bm, TN_model_tot_gap)

# transform predictions
TN_predict_tot_gap_transform <- TN_predict_tot_gap
values(TN_predict_tot_gap_transform) <- 1 / (1 + exp(-values(TN_predict_tot_gap))) 

# export
writeRaster(TN_predict_tot_gap_transform, filename="data created/TN_predict_gap.tif", overwrite=TRUE)
writeRaster(TN_predict_tot_gap, filename="data created/TN_predict_gap1.tif", overwrite=TRUE)

```

# TSS


## GM
```{r}

# extract prediction data from sample sites
valid_pts2 <- terra::extract(GM_predict_tot_gap_transform, data.frame(GM_gap_env$MEAN_X, GM_gap_env$MEAN_Y)) |> 
  cbind(data.frame(MEAN_X = GM_gap_env$MEAN_X, MEAN_Y = GM_gap_env$MEAN_Y))

GM_gap_env_pred <- left_join(GM_gap_env, valid_pts2) |> 
  mutate(gap_pred = ifelse(lyr1 >= 0.5, 1, 0)) |> 
  na.omit()

GM_gap_env_pred$gap <- factor(GM_gap_env_pred$gap, levels = c(0,1))
GM_gap_env_pred$gap_pred <- factor(GM_gap_env_pred$gap_pred, levels = c(0,1))


# create confusion matrix comparing classifications to predictions
CM_GM <- confusionMatrix(GM_gap_env_pred$gap_pred, GM_gap_env_pred$gap)
```

## TN
```{r}

# extract prediction data from sample sites
valid_pts2 <- terra::extract(TN_predict_tot_gap_transform, data.frame(TN_gap_env$MEAN_X, TN_gap_env$MEAN_Y)) |> 
  cbind(data.frame(MEAN_X = TN_gap_env$MEAN_X, MEAN_Y = TN_gap_env$MEAN_Y))

TN_gap_env_pred <- left_join(TN_gap_env, valid_pts2) |> 
  mutate(gap_pred = ifelse(lyr1 >= 0.5, 1, 0)) |> 
  na.omit()

TN_gap_env_pred$gap <- factor(TN_gap_env_pred$gap, levels = c(0,1))
TN_gap_env_pred$gap_pred <- factor(TN_gap_env_pred$gap_pred, levels = c(0,1))


# create confusion matrix comparing classifications to predictions
CM_TN <- confusionMatrix(TN_gap_env_pred$gap_pred, TN_gap_env_pred$gap)

```

## together
```{r}
gap_env_pred <- rbind(TN_gap_env_pred, GM_gap_env_pred)

CM <- confusionMatrix(gap_env_pred$gap_pred, gap_env_pred$gap)
```

# loop through cutoffs
```{r}
# function to loop through gap cut off 
## then use to graph 

cutoff_gapsTSS <- function(cutoff) {
  # extract prediction data from sample sites
  valid_pts2 <- terra::extract(GM_predict_tot_gap_transform, data.frame(GM_gap_env$MEAN_X, GM_gap_env$MEAN_Y)) |> 
    cbind(data.frame(MEAN_X = GM_gap_env$MEAN_X, MEAN_Y = GM_gap_env$MEAN_Y))
  
  GM_gap_env_pred <- left_join(GM_gap_env, valid_pts2) |> 
    mutate(gap_pred = ifelse(lyr1 >= cutoff, 1, 0)) |> 
    na.omit()
  
  GM_gap_env_pred$gap <- factor(GM_gap_env_pred$gap, levels = c(0,1))
  GM_gap_env_pred$gap_pred <- factor(GM_gap_env_pred$gap_pred, levels = c(0,1))
  
  
  # create confusion matrix comparing classifications to predictions
  CM_GM <- confusionMatrix(GM_gap_env_pred$gap_pred, GM_gap_env_pred$gap)
  
  # calculate TSS
  CM_OA_GM <- (CM_GM$table[1,1]/(CM_GM$table[1,1] + CM_GM$table[2,1])) + (CM_GM$table[2,2]/(CM_GM$table[2,2] + CM_GM$table[1,2])) -1
  
  # extract prediction data from sample sites
  valid_pts2 <- terra::extract(TN_predict_tot_gap_transform, data.frame(TN_gap_env$MEAN_X, TN_gap_env$MEAN_Y)) |> 
    cbind(data.frame(MEAN_X = TN_gap_env$MEAN_X, MEAN_Y = TN_gap_env$MEAN_Y))
  
  TN_gap_env_pred <- left_join(TN_gap_env, valid_pts2) |> 
    mutate(gap_pred = ifelse(lyr1 >= cutoff, 1, 0)) |> 
    na.omit()
  
  TN_gap_env_pred$gap <- factor(TN_gap_env_pred$gap, levels = c(0,1))
  TN_gap_env_pred$gap_pred <- factor(TN_gap_env_pred$gap_pred, levels = c(0,1))
  
  
  # create confusion matrix comparing classifications to predictions
  CM_TN <- confusionMatrix(TN_gap_env_pred$gap_pred, TN_gap_env_pred$gap)
  
  # calculate TSS
  CM_OA_TN = (CM_TN$table[1,1]/(CM_TN$table[1,1] + CM_TN$table[2,1])) + (CM_TN$table[2,2]/(CM_TN$table[2,2] + CM_TN$table[1,2])) -1
  
  # both parks together
  gap_env_pred <- rbind(TN_gap_env_pred, GM_gap_env_pred)
  
  CM <- confusionMatrix(gap_env_pred$gap_pred, gap_env_pred$gap)
  
  # calculate TSS  
  CM_OA = (CM$table[1,1]/(CM$table[1,1] + CM$table[2,1])) + (CM$table[2,2]/(CM$table[2,2] + CM$table[1,2])) -1
  
  
  return(list(CM_OA, CM_OA_TN, CM_OA_GM))
}


TSS_CM <- data.frame(t(sapply(cutoff_list, cutoff_gapsTSS))) |> 
  mutate(
    both = unlist(X1),
    TN = unlist(X2),
    GM = unlist(X3)
  ) |> 
  cbind(cutoff_list) 

ggplot(TSS_CM) +
  geom_point(aes(cutoff_list, both)) +
  geom_point(aes(cutoff_list, TN), colour = "red") +
  geom_point(aes(cutoff_list, GM), colour = "blue") +
  ylab("TSS")
```
