==> code_for_stats/code availability/11) c error propagation.Rmd <==
---
title: "11) c error propagation"
author: "Emmerson Wilson"
date: "2024-05-08"
output: html_document
---
```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  terra)

```

```{r}
# change this to whichever folder you downloaded the code to
setwd("~/Documents/Master's/C_forecasting_NL/code_for_stats/code availability")

# import glm function
source("2) glm function.R")
```


# import training datasets
```{r}
# made in 5) c lm total.Rmd
GM_c_env <- read_xlsx("data created/c_env_sites_scale_GM.xlsx")
TN_c_env <- read_xlsx("data created/c_env_sites_scale_TN.xlsx")
```

# bring scaled rasters
```{r}
GM_env_raster_masked_scale_bm <- rast("data created/GM_env_raster_masked_scale.tif")
TN_env_raster_masked_scale_bm <- rast("data created/TN_env_raster_masked_scale.tif")
```


# GM
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable

## model
```{r}
# list of variables in best model from 5) c lm total.Rmd
GM_var <- list("FHT", "EVIamp")

# run best model
GM_model_tot <- lm_glm(GM_var, df = GM_c_env, c = "c_m2_site")
summary(GM_model_tot)
```

```{r}
GM_predict_tot1 <- terra::predict(GM_env_raster_masked_scale_bm, GM_model_tot, se.fit = T)
GM_predict_tot <- exp(GM_predict_tot1)

```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
GM_predict_tot_se1 <- GM_predict_tot1$se.fit # untransformed
GM_predict_tot_se <- exp(GM_predict_tot_se1)

# export
writeRaster(GM_predict_tot_se, filename="data created/GM_predict_tot_se.tif", overwrite=TRUE)
# export
writeRaster(GM_predict_tot_se1, filename="data created/GM_predict_tot_se1.tif", overwrite=TRUE)
```

## calculate CI from se of fit
```{r}
GM_CI_upper_95 <- exp(GM_predict_tot1$fit + 1.96*GM_predict_tot_se1)
GM_CI_lower_95 <- exp(GM_predict_tot1$fit - 1.96*GM_predict_tot_se1)
```

## export CI
```{r}
writeRaster(GM_CI_upper_95, filename="data created/GM_CI_upper_95.tif", overwrite=TRUE)
writeRaster(GM_CI_lower_95, filename="data created/GM_CI_lower_95.tif", overwrite=TRUE)
```


## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
GM_se.PI1 <- sqrt(GM_predict_tot_se1$se.fit ^ 2 + GM_predict_tot1$residual.scale ^ 2) # untransformed
GM_se.PI <- exp(GM_se.PI1)

# export
writeRaster(GM_se.PI, filename="data created/GM_predict_tot_pse.tif", overwrite=TRUE)

# export
writeRaster(GM_se.PI1, filename="data created/GM_predict_tot_pse1.tif", overwrite=TRUE)
```

## calculate PI from se fo residual variance
```{r}
GM_PI_upper_95 <- exp(GM_predict_tot1$fit + 1.96*GM_se.PI1)
GM_PI_lower_95 <- exp(GM_predict_tot1$fit - 1.96*GM_se.PI1)
```

## export PI
```{r}
writeRaster(GM_PI_upper_95, filename="data created/GM_PI_upper_95.tif", overwrite=TRUE)
writeRaster(GM_PI_lower_95, filename="data created/GM_PI_lower_95.tif", overwrite=TRUE)
```

# TN
## make predictions using top model and stacked environmental rasters, transformed onto scale of dependant variable

## mdoel
```{r}
# list of variables in best model from 5) c lm total.Rmd
TN_var <- list("FHT", "EVIamp")

# run best model
TN_model_tot <- lm_glm(TN_var, df = subset(TN_c_env), c = "c_m2_site")
summary(TN_model_tot)
```

## make predictions
```{r}
TN_predict_tot1 <- terra::predict(TN_env_raster_masked_scale_bm, TN_model_tot, se.fit = T)
TN_predict_tot <- exp(TN_predict_tot1)
```

## get standard errors of the predicted mean from predict() function, transform to scale of dependent variable 
```{r}
TN_predict_tot_se1 <- TN_predict_tot1$se.fit
TN_predict_tot_se <- exp(TN_predict_tot_se1)

# export
writeRaster(TN_predict_tot_se, filename="data created/TN_predict_tot_se.tif", overwrite=TRUE)
writeRaster(TN_predict_tot_se1, filename="data created/TN_predict_tot_se1.tif", overwrite=TRUE)

```

## calculate CI from se of fit
```{r}
TN_CI_upper_95 <- exp(TN_predict_tot1$fit + 1.96*TN_predict_tot_se1)
TN_CI_lower_95 <- exp(TN_predict_tot1$fit - 1.96*TN_predict_tot_se1)
```

## export CI
```{r}
writeRaster(TN_CI_upper_95, filename="data created/TN_CI_upper_95.tif", overwrite=TRUE)
writeRaster(TN_CI_lower_95, filename="data created/TN_CI_lower_95.tif", overwrite=TRUE)
```

## get standard errors from residual variance (of prediction interval), transfomr to scale of dependant variable
```{r}
TN_se.PI1 <- sqrt(TN_predict_tot_se1$se.fit ^ 2 + TN_predict_tot1$residual.scale ^ 2)
TN_se.PI <- exp(TN_se.PI1)

#export
writeRaster(TN_se.PI, filename="data created/TN_predict_tot_pse.tif", overwrite=TRUE)
writeRaster(TN_se.PI1, filename="data created/TN_predict_tot_pse1.tif", overwrite=TRUE)
```

## calculate PI from se fo residual variance
```{r}
TN_PI_upper_95 <- exp(TN_predict_tot1$fit + 1.96*TN_se.PI1)
TN_PI_lower_95 <- exp(TN_predict_tot1$fit - 1.96*TN_se.PI1)
```

## export PI
```{r}
writeRaster(TN_PI_upper_95, filename="data created/TN_PI_upper_95.tif", overwrite=TRUE)
writeRaster(TN_PI_lower_95, filename="data created/TN_PI_lower_95.tif", overwrite=TRUE)
```
