==> code_for_stats/code availability/18) without GM 35.Rmd <==
---
title: "19) without GM 35"
author: "Emmerson Wilson"
date: "2024-07-23"
output: html_document
---

```{r}
pacman::p_load(
  dplyr,
  tidyverse,
  readxl,
  openxlsx,
  cv,
  lme4,
  AICcmodavg,
  rcompanion,
  boot)
  
```

```{r}
# change this to whichever folder you downloaded the code to
setwd("/Users/emmersonEmmerson/Documents/Master's/C_forecasting_NL/code_for_stats/clean code")
```

### import datasets
```{r}
c_env_sites_scale_GM35 <- read_xlsx("data created/c_env_sites_scale_GM.xlsx") %>% 
  filter(unique_id != "GM_35_2023")
```

# glm function
extremely useful for univariate models
```{r}
source("4) glm function.R")
```


# GM

### null

```{r}
#run the model
GM35_lm_null <- lm_glm(c_env_sites_scale_GM35, "c_m2_site", "1")
```

```{r}
# get the summary of the model into a dataframe
GM35_lm_null_summary <- as.data.frame(summary(GM35_lm_null)$coefficients) %>% 
  mutate(model = "GM35_lm_null") 
GM35_lm_null_summary <- GM35_lm_null_summary %>% 
  add_column("predictor" = rownames(GM35_lm_null_summary))

# reorder columns so model name first
GM35_lm_null_summary <- GM35_lm_null_summary[,c(5,6,1:4)]
GM35_lm_null_summary
```

### full

```{r}
# make list of all variables to be included in full model
full_var <- list("FHT", "EVIamp", "EVImed", "ELE", "SLO", "ASP", "LGS", "Moose_new")

# run the model
GM35_lm_full <- lm_glm(c_env_sites_scale_GM35, "c_m2_site", full_var)
```

```{r}
# add summary of model to dataframe
GM35_lm_full_summary <- as.data.frame(summary(GM35_lm_full)$coefficients) %>% 
  mutate(model = "GM35_lm_full") 
GM35_lm_full_summary <- GM35_lm_full_summary %>% 
  add_column("predictor" = rownames(GM35_lm_full_summary))

GM35_lm_full_summary <- GM35_lm_full_summary[,c(5,6,1:4)]
GM35_lm_full_summary

```

### univariate

```{r}
# run model with each predictor variable on own

GM35_lm_uni <- lapply(full_var, lm_glm, df = c_env_sites_scale_GM35, c = "c_m2_site")

names(GM35_lm_uni) <- full_var # make sure names of objects in list consistent
```

```{r}
GM35_lm_uni_summary <- lapply(GM35_lm_uni, summary)

GM35_lm_uni_summary

GM35_lm_uni_AIC <- as.data.frame(t(as.data.frame(lapply(GM35_lm_uni, AICc)))) 
GM35_lm_uni_AIC$model <- rownames(GM35_lm_uni_AIC)

# get summary of each univariate model, add them to a single dataframe with model name and variables in own columns

GM35_uni_summary_all <- Reduce(full_join,lapply(mapply(cbind, lapply(lapply(GM35_lm_uni_summary, coef), as.data.frame), "model"=names(GM35_lm_uni_summary), SIMPLIFY=F), rownames_to_column, var = "predictor")) 
GM35_uni_summary_all <- GM35_uni_summary_all[,c(5,6,1:4)]
GM35_uni_summary_all
```

### reduced
```{r}
# find possible reduced model to add to 
# need to input full model not through function 
GM35_lm_full2 <- glm(formula = c_m2_site ~ FHT + EVIamp + EVImed + ELE + SLO + ASP + LGS + Moose_new, data = c_env_sites_scale_GM35, family = Gamma(link = "log")) 

GM35_lm_red <- step(GM35_lm_full2, direction = "backward", trace = F)

```

```{r}
GM35_lm_red_summary <- as.data.frame(summary(GM35_lm_red)$coefficients) %>% 
  mutate(model = "GM35_lm_red") 
GM35_lm_red_summary <- GM35_lm_red_summary %>% 
  add_column("predictor" = rownames(GM35_lm_red_summary))

GM35_lm_red_summary <- GM35_lm_red_summary[,c(5,6,1:4)]
GM35_lm_red_summary

GM35_lm_red2 <- glm(formula = c_m2_site ~ FHT + EVImed + EVIamp, data = c_env_sites_scale_GM35, family = Gamma(link = "log")) 
GM35_lm_red3 <- glm(formula = c_m2_site ~ FHT + EVIamp, data = c_env_sites_scale_GM35, family = Gamma(link = "log")) 
GM35_lm_red4 <- glm(formula = c_m2_site ~ EVImed + EVIamp, data = c_env_sites_scale_GM35, family = Gamma(link = "log")) 

GM35_red_models_r <- list()
GM35_red_models_r[[names(GM35_lm_uni)[1]]] <- GM35_lm_uni[[1]]
GM35_red_models_r[[names(GM35_lm_uni)[2]]] <- GM35_lm_uni[[2]]
GM35_red_models_r[[names(GM35_lm_uni)[3]]] <- GM35_lm_uni[[3]]
GM35_red_models_r[["GM35_lm_red"]] <- GM35_lm_red
GM35_red_models_r[["GM35_lm_red2"]] <- GM35_lm_red2
GM35_red_models_r[["GM35_lm_red3"]] <- GM35_lm_red3
GM35_red_models_r[["GM35_lm_red4"]] <- GM35_lm_red4
Modnames_GM35_red <- paste(names(GM35_red_models_r))

# get table
aic_GM35red <- aictab(cand.set = GM35_red_models_r, modnames = Modnames_GM35_red, sort = TRUE) 
```

# cross validation and glm functions
```{r}

# glm cv function
cv_cv.glm2 <- function(data, vars,  k) {
c_Output <- NULL

#Do it 100 times
for(x in seq(1,100,1)){
  
  #Step 1 - Re-shuffle the response
  #Reshuffle stoich data and place in new file
  c_Shuf <- transform(data, c_m2_site = sample(c_m2_site))
  
  #Step 2
  #Run the model
  #Need to do glm because boot works on glm. Normal distribution so same as lm.
  c_Rand <- lm_glm(vars, df = c_Shuf, c = "c_m2_site")
  
  #Step 3
  #Run cross validation
  cv <-cv.glm(data=c_Shuf, glmfit=c_Rand, K=k)
  
  
  #Step 4
  #Collect the cv values for each model. For each iteration x save the cv.
  c_Output <- rbind(c_Output, data.frame(Iteration = x, CV = cv$delta[2]))
  
  print (x)
}

#Step 5
#Sort random set to calculate 90% CI for cv$delta[2]
c_Output_Sort <- c_Output[order(c_Output$CV),]

t.test(c_Output_Sort$CV)
}
```

## GM35

### top models from AIC
```{r}
# list of variables in best models from 5) c lm total.Rmd
# run them
var_GM35_lm_red <- list("FHT", "EVImed")
GM35_c_red <- lm_glm(var_GM35_lm_red, df = c_env_sites_scale_GM35, c = "c_m2_site")

var_GM35_lm_red2 <- list("FHT", "EVIamp", "EVImed")
GM35_c_red2 <- lm_glm(var_GM35_lm_red2, df = c_env_sites_scale_GM35, c = "c_m2_site")

var_GM35_lm_red3 <- list("FHT", "EVIamp")
GM35_c_red3 <- lm_glm(var_GM35_lm_red3, df = c_env_sites_scale_GM35, c = "c_m2_site")

var_GM35_lm_FHT <- list("FHT")
GM35_c_FHT <- lm_glm(var_GM35_lm_FHT, df = c_env_sites_scale_GM35, c = "c_m2_site")

GM35_var_lists <- list(var_GM35_lm_red, var_GM35_lm_red2, var_GM35_lm_red3, var_GM35_lm_FHT)
GM35_c_top_models <- list(GM35_c_red, GM35_c_red2, GM35_c_red3, GM35_c_FHT)

# do cross validation on models themselves
k <- 9
GM35_cv_Output <- NULL

for(x in seq(1,100,1)){
GM35_c_cv <- lapply(GM35_c_top_models, cv.glm, data=c_env_sites_scale_GM35, K = k)
names(GM35_c_cv) <- GM35_var_lists

GM35_c_cv_delta <- lapply(GM35_c_cv, `[`, c('delta'))

GM35_c_cv_delta <- do.call(rbind.data.frame, GM35_c_cv_delta)
GM35_c_cv_delta$names <- rownames(GM35_c_cv_delta)

GM35_cv_Output <- rbind(GM35_cv_Output, data.frame(Iteration = x, model = GM35_c_cv_delta$names, delta = GM35_c_cv_delta$delta))
}

GM35_avg_cv <- GM35_cv_Output %>% 
  group_by(model) %>% 
  summarize(mean_delta1 = mean(delta)) 
  
GM35_cv_Output_delta1_2 <- GM35_cv_Output %>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("delta_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = delta, names_from = n)
```

### cross validation of random 
```{r}
GM35_c_cvrand <- lapply(GM35_var_lists, FUN = cv_cv.glm2, d = c_env_sites_scale_GM35, k = k)
names(GM35_c_cvrand) <- GM35_var_lists

GM35_c_cvrand_95 <- lapply(GM35_c_cvrand, `[`, c('conf.int'))

GM35_c_cvrand_95 <- do.call(rbind.data.frame, GM35_c_cvrand_95)
GM35_c_cvrand_95$model <- rownames(GM35_c_cvrand_95)

GM35_c_cvrand_95_1_2 <- GM35_c_cvrand_95%>% 
  mutate(model_2 = substr(model,1,nchar(model)-2)) %>% 
  mutate(n = paste0("ci95_",str_sub(model, -1, -1)))  %>% 
  select(!2) %>% 
  pivot_wider(values_from = conf.int, names_from = n)
```

## put random with cv
```{r}
GM35_val_cv <- GM35_cv_Output_delta1_2 %>% 
  full_join(GM35_c_cvrand_95_1_2) %>% 
  mutate(diff = ci95_1 - delta_1) %>% 
  mutate(validate = ifelse(diff<0.0, "no", "yes"))
GM35_nval_cv <- GM35_val_cv %>% 
  group_by(model_2) %>% 
  summarize(n_y = sum(validate == "yes"), n_n = sum(validate == "no"))
```
